POSTリクエストにおけるURLRequestオブジェクト作成の責務をRepository層から外した
